backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
celery==5.6.1
redis==7.1.0
django-celery-results==2.6.0
django-celery-beat>=2.7.0
mysqlclient==2.2.8
cryptography==46.0.5
requests>=2.31.0
reportlab>=4.0.0
WeasyPrint>=63.0

dnspython>=2.7.0

